gene synchro complete (not tested)

+++ bugs/entities/Gamete.cpp
@@ -50,7 +50,7 @@ void Gamete::onCollision(PhysicsBody* pOther, float impulse) {
 	Gamete *other = (Gamete*)pOther->userPointer_;
 	if (other->isZombie())
 		return;
-	if (abs(other->getChromosome().genes.size-chromosome_.genes.size) > WorldConst::MaxGenomeLengthDifference)
+	if (abs(other->getChromosome().genes.size()-chromosome_.genes.size()) > (uint)WorldConst::MaxGenomeLengthDifference)
 		return;
 	Genome g;
 	g.first = chromosome_;
@@ -80,7 +80,7 @@ void Gamete::update(float dt) {
 			// found another gamete, let's attract each other
 			// p.s.: only attract and fuse if they don't differ by more than N in the number of genes - N is the same as the new gene recorded history
 			Gamete* other = (Gamete*)((PhysicsBody*)b->GetUserData())->userPointer_;
-			if (abs(other->getChromosome().genes.size-chromosome_.genes.size) > WorldConst::MaxGenomeLengthDifference)
+			if (abs(other->getChromosome().genes.size()-chromosome_.genes.size()) > (uint)WorldConst::MaxGenomeLengthDifference)
 				continue;
 			b2Vec2 force = body_.b2Body_->GetPosition() - b->GetPosition();
 			float distance = force.Normalize();
